+++ cmd/helm-unittest/helm_unittest.go
@@ -14,6 +14,8 @@ import (
 // testOptions stores options setup by user in command line
 type testOptions struct {
 	useHelmV3      bool
+	useFailfast    bool
+	useStrict      bool
 	colored        bool
 	updateSnapshot bool
 	withSubChart   bool
@@ -72,6 +74,8 @@ details about how to write tests.
 			Formatter:      formatter,
 			UpdateSnapshot: testConfig.updateSnapshot,
 			WithSubChart:   testConfig.withSubChart,
+			Strict:         testConfig.useStrict,
+			Failfast:       testConfig.useFailfast,
 			TestFiles:      testConfig.testFiles,
 			OutputFile:     testConfig.outputFile,
 		}
@@ -103,6 +107,11 @@ func init() {
 		"enforce printing colored output even stdout is not a tty. Set to false to disable color",
 	)
 
+	cmd.PersistentFlags().BoolVar(
+		&testConfig.useStrict, "strict", false,
+		"strict parse the testsuites",
+	)
+
 	defaultFilePattern := filepath.Join("tests", "*_test.yaml")
 	cmd.PersistentFlags().StringArrayVarP(
 		&testConfig.testFiles, "file", "f", []string{defaultFilePattern},
@@ -131,6 +140,11 @@ func init() {
 
 	cmd.PersistentFlags().BoolVarP(
 		&testConfig.useHelmV3, "helm3", "3", false,
-		"parse helm charts as helm3 charts.",
+		"parse helm charts as helm3 charts",
+	)
+
+	cmd.PersistentFlags().BoolVarP(
+		&testConfig.useFailfast, "failfast", "q", false,
+		"direct quit testing, when a test is failed",
 	)
 }