+++ pkg/unittest/test_runner.go
@@ -60,6 +60,8 @@ type TestRunner struct {
 	Formatter        formatter.Formatter
 	UpdateSnapshot   bool
 	WithSubChart     bool
+	Strict           bool
+	Failfast         bool
 	TestFiles        []string
 	OutputFile       string
 	suiteCounting    testUnitCountingWithSnapshotFailed
@@ -79,6 +81,9 @@ func (tr *TestRunner) RunV2(ChartPaths []string) bool {
 			tr.printErroredChartHeader(err)
 			tr.countChart(false, err)
 			allPassed = false
+			if tr.Failfast {
+				break
+			}
 			continue
 		}
 		chartRoute := chart.Metadata.Name
@@ -87,6 +92,9 @@ func (tr *TestRunner) RunV2(ChartPaths []string) bool {
 			tr.printErroredChartHeader(err)
 			tr.countChart(false, err)
 			allPassed = false
+			if tr.Failfast {
+				break
+			}
 			continue
 		}
 
@@ -115,6 +123,9 @@ func (tr *TestRunner) RunV3(ChartPaths []string) bool {
 			tr.printErroredChartHeader(err)
 			tr.countChart(false, err)
 			allPassed = false
+			if tr.Failfast {
+				break
+			}
 			continue
 		}
 		chartRoute := chart.Name()
@@ -123,6 +134,9 @@ func (tr *TestRunner) RunV3(ChartPaths []string) bool {
 			tr.printErroredChartHeader(err)
 			tr.countChart(false, err)
 			allPassed = false
+			if tr.Failfast {
+				break
+			}
 			continue
 		}
 
@@ -155,12 +169,13 @@ func (tr *TestRunner) getTestSuites(chartPath, chartRoute string) ([]*TestSuite,
 
 	resultSuites := make([]*TestSuite, 0, len(filesSet))
 	for file := range filesSet {
-		suite, err := ParseTestSuiteFile(file, chartRoute)
+		suite, err := ParseTestSuiteFile(file, chartRoute, tr.Strict)
 		if err != nil {
 			tr.handleSuiteResult(&results.TestSuiteResult{
 				FilePath:  file,
 				ExecError: err,
 			})
+			return nil, err
 		}
 		resultSuites = append(resultSuites, suite)
 	}
@@ -230,12 +245,16 @@ func (tr *TestRunner) runV2SuitesOfChart(suites []*TestSuite, chart *v2chart.Cha
 			continue
 		}
 
-		result := suite.RunV2(chart, snapshotCache, &results.TestSuiteResult{})
+		result := suite.RunV2(chart, snapshotCache, tr.Failfast, &results.TestSuiteResult{})
 		chartPassed = chartPassed && result.Passed
 		tr.handleSuiteResult(result)
 		tr.testResults = append(tr.testResults, result)
 
 		snapshotCache.StoreToFileIfNeeded()
+
+		if !chartPassed && tr.Failfast {
+			break
+		}
 	}
 
 	return chartPassed
@@ -255,12 +274,16 @@ func (tr *TestRunner) runV3SuitesOfChart(suites []*TestSuite, chart *v3chart.Cha
 			continue
 		}
 
-		result := suite.RunV3(chart, snapshotCache, &results.TestSuiteResult{})
+		result := suite.RunV3(chart, snapshotCache, tr.Failfast, &results.TestSuiteResult{})
 		chartPassed = chartPassed && result.Passed
 		tr.handleSuiteResult(result)
 		tr.testResults = append(tr.testResults, result)
 
 		snapshotCache.StoreToFileIfNeeded()
+
+		if !chartPassed && tr.Failfast {
+			break
+		}
 	}
 
 	return chartPassed